refactor: create useProduct hooks
